+++ src/config/dictionary/parsing/actions.mjs
@@ -7,6 +7,7 @@ export const PARSING_ACTIONS = {
     "Form of the Beast: Bite",
     "Form of the Beast: Claws",
     "Form of the Beast: Tail",
+    "Cat's Claws",
   ],
   KEEP_ACTIONS_STARTSWITH: [
     "Metamagic:",
@@ -51,7 +52,7 @@ export const PARSING_ATTACK_ACTIONS = {
     "Talon",
     "Trunk",
     "Claws",
-    "Cat's Claws",
+    // "Cat's Claws",
     "Fangs",
     "Form of the Beast: Bite",
     "Form of the Beast: Claws",
@@ -65,5 +66,6 @@ export const PARSING_ATTACK_ACTIONS = {
     "Psychic Blades: Bonus Attack (STR)",
     "Psychic Blades",
     "Psychic Blade",
+    "Cat's Claws",
   ],
 };